update styling in admin update class page
