+++ src/pages/Admin/AdminUpdateClass/styles.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const iconStyles = {
+  fontSize: `1.3rem`,
+  opacity: `50%`,
+  cursor: `pointer`,
+  transition: `all ease 0.1s`,
+  '&:hover': {
+    opacity: '100%',
+  },
+};
+
+export const StyledDateString = styled.span`
+  font-weight: 100;
+  font-style: italic;
+`;
